refactor(mvp): remove other stuff
