--- src/layouts/EditPage/BlocksEditPage/ImageMenuBar.tsx
@@ -1,158 +1,0 @@
-import { Divider, HStack } from "@chakra-ui/react"
-import { Editor } from "@tiptap/react"
-import { Fragment } from "react"
-
-import { useEditorModal } from "contexts/EditorModalContext"
-
-import { MenuItem } from "../../components/Editor/components/MenuItem"
-
-export const ImageMenuBar = ({ editor }: { editor: Editor }) => {
-  const { showModal } = useEditorModal()
-
-  const items = [
-    {
-      icon: "bold",
-      title: "Bold",
-      action: () => editor.chain().focus().toggleBold().run(),
-      isActive: () => editor.isActive("bold"),
-    },
-    {
-      icon: "italic",
-      title: "Italic",
-      action: () => editor.chain().focus().toggleItalic().run(),
-      isActive: () => editor.isActive("italic"),
-    },
-    {
-      icon: "strikethrough",
-      title: "Strike",
-      action: () => editor.chain().focus().toggleStrike().run(),
-      isActive: () => editor.isActive("strike"),
-    },
-    {
-      icon: "code-view",
-      title: "Code",
-      action: () => editor.chain().focus().toggleCode().run(),
-      isActive: () => editor.isActive("code"),
-    },
-    {
-      icon: "mark-pen-line",
-      title: "Highlight",
-      action: () => editor.chain().focus().toggleHighlight().run(),
-      isActive: () => editor.isActive("highlight"),
-    },
-    {
-      type: "divider",
-    },
-    {
-      icon: "h-1",
-      title: "Heading 1",
-      action: () => editor.chain().focus().toggleHeading({ level: 1 }).run(),
-      isActive: () => editor.isActive("heading", { level: 1 }),
-    },
-    {
-      icon: "h-2",
-      title: "Heading 2",
-      action: () => editor.chain().focus().toggleHeading({ level: 2 }).run(),
-      isActive: () => editor.isActive("heading", { level: 2 }),
-    },
-    {
-      icon: "h-3",
-      title: "Heading 3",
-      action: () => editor.chain().focus().toggleHeading({ level: 3 }).run(),
-      isActive: () => editor.isActive("heading", { level: 3 }),
-    },
-    {
-      icon: "paragraph",
-      title: "Paragraph",
-      action: () => editor.chain().focus().setParagraph().run(),
-      isActive: () => editor.isActive("paragraph"),
-    },
-    {
-      type: "divider",
-    },
-    {
-      icon: "double-quotes-l",
-      title: "Blockquote",
-      action: () => editor.chain().focus().toggleBlockquote().run(),
-      isActive: () => editor.isActive("blockquote"),
-    },
-    {
-      icon: "separator",
-      title: "Horizontal Rule",
-      action: () => editor.chain().focus().setHorizontalRule().run(),
-    },
-    {
-      type: "divider",
-    },
-    {
-      icon: "text-wrap",
-      title: "Hard Break",
-      action: () => editor.chain().focus().setHardBreak().run(),
-    },
-    {
-      icon: "format-clear",
-      title: "Clear Format",
-      action: () => editor.chain().focus().clearNodes().unsetAllMarks().run(),
-    },
-    {
-      type: "divider",
-    },
-    {
-      icon: "arrow-go-back-line",
-      title: "Undo",
-      action: () => editor.chain().focus().undo().run(),
-    },
-    {
-      icon: "arrow-go-forward-line",
-      title: "Redo",
-      action: () => editor.chain().focus().redo().run(),
-    },
-    {
-      icon: "file-image-line",
-      title: "Add image",
-      action: () =>
-        editor
-          .chain()
-          .focus()
-          .setImage({
-            src:
-              "https://upload.wikimedia.org/wikipedia/commons/thumb/3/3f/Placeholder_view_vector.svg/1362px-Placeholder_view_vector.svg.png",
-          })
-          .run(),
-    },
-    {
-      icon: "links-line",
-      title: "Add link",
-      isActive: () => editor.isActive("link"),
-      action: () => editor.chain().focus().unsetLink().run(),
-    },
-  ]
-
-  return (
-    <HStack
-      bgColor="gray.50"
-      flex="0 0 auto"
-      flexWrap="wrap"
-      p="0.25rem"
-      borderBottom="1px solid"
-      borderColor="base.divider.strong"
-      borderTopRadius="0.25rem"
-    >
-      {items.map((item, index) => (
-        // eslint-disable-next-line react/no-array-index-key
-        <Fragment key={index}>
-          {item.type === "divider" ? (
-            <Divider
-              orientation="vertical"
-              border="px solid"
-              borderColor="base.divider.strong"
-              h="1.25rem"
-            />
-          ) : (
-            <MenuItem {...item} />
-          )}
-        </Fragment>
-      ))}
-    </HStack>
-  )
-}
