--- src/layouts/EditPage/BlocksEditPage/index.ts
@@ -1,1 +1,0 @@
-export * from "./BlocksEditPage"
