--- src/layouts/EditPage/TiptapBlocksEditPage/index.ts
@@ -1,1 +1,0 @@
-export * from "./TiptapBlocksEditPage"
